+++ quiz-app/documentation/change_log.md
@@ -16,3 +16,7 @@
 - Added and tested new features for user playing the quiz for QuestionService and QuestionRepository classes (checking answers, adding points, getting points)
 - User can see the questions one by one in random order and answer these, correct answer will increase the quiz points by +1 and incorrect does not increase, this is saved to the database.
 - User can see the total points of the quiz once all of the questions are answered
+  
+## Week5
+- User can see the current highscore before starting the quiz, which is replaced by quiz points if those will be higher.
+- Added highscore column to the database and added the related logic to question services and database changes to question repositories.